decode Healthcheck.Test using DecodeMapstructure
